DWT and Showcase updated to support the latest Dart SDK (0.8.7)

+++ lib/src/i18n/number_format.dart
@@ -340,7 +340,7 @@ class NumberFormat {
    */
   String format(num number) {
     if (number is int) {
-      return formatInt(number as int);
+      return formatInt(number);
     } else {
       return formatDouble(number as double);
     }
